requirements.txt
# Core Framework
google-adk
litellm
python-dotenv
chromadb

# API & Web UI
fastapi
uvicorn
sse-starlette
python-multipart

# Data Validation & Templating
pydantic
jinja2

# Infrastructure Adapters (Role 2 Mandates)
asyncpg
redis[hiredis]

# Testing & Observability (Test Lead Mandates)
pytest
pytest-asyncio
arize-phoenix
testcontainers[postgres]
testcontainers[redis]
toxiproxy-python
vcrpy
psycopg2-binary
pgvector
